+++ handlers/today_button.py
@@ -15,30 +15,32 @@ router_today = Router()
 async def get_today(message: Message, state: FSMContext):
     await support_function.user_tracking.where_who(where=message.text,
                                                    state=state)
-    user_data = await state.get_data()
-    user_sn = user_data['user_second_name']
-    if user_sn in ('Василевский', 'Крусcер'):
-        msg = await message.answer(text='Для вас будет использоваться график Быковой Ангелины!')
-        await support_function.delete_pre_message.del_pre_message(chat_id=msg.chat.id,
-                                                                  message_id=msg.message_id,
-                                                                  state=state)
-        # возможно будет изменения в if_work и [0] не будет
-        await message.delete()
-        msg2 = await message.answer(text='Идет загрузка данных...')
-        msg1 = await message.answer(text=ResultPrint('Быкова').today_button()[0],
-                                    reply_markup=main_kb)
-        await state.update_data(whitch_kb_was='main_kb')
-        await msg2.delete()
-        await support_function.delete_pre_message.del_pre_message(chat_id=msg1.chat.id,
-                                                                  message_id=msg1.message_id,
-                                                                  state=state)
-    else:
-        await message.delete()
-        msg1 = await message.answer(text='Идет загрузка данных...')
-        msg = await message.answer(text=ResultPrint(user_sn).today_button()[0],
-                                   reply_markup=main_kb)
-        await state.update_data(whitch_kb_was='main_kb')
-        await msg1.delete()
-        await support_function.delete_pre_message.del_pre_message(chat_id=msg.chat.id,
-                                                                  message_id=msg.message_id,
-                                                                  state=state)
+    if await support_function.login_test.log_test(message=message,
+                                                  state=state):
+        user_data = await state.get_data()
+        user_sn = user_data['user_second_name']
+        if user_sn in ('Василевский', 'Крусcер'):
+            msg = await message.answer(text='Для вас будет использоваться график Быковой Ангелины!')
+            await support_function.delete_pre_message.del_pre_message(chat_id=msg.chat.id,
+                                                                      message_id=msg.message_id,
+                                                                      state=state)
+            # возможно будет изменения в if_work и [0] не будет
+            await message.delete()
+            msg2 = await message.answer(text='Идет загрузка данных...')
+            msg1 = await message.answer(text=ResultPrint('Быкова').today_button()[0],
+                                        reply_markup=main_kb)
+            await state.update_data(whitch_kb_was='main_kb')
+            await msg2.delete()
+            await support_function.delete_pre_message.del_pre_message(chat_id=msg1.chat.id,
+                                                                      message_id=msg1.message_id,
+                                                                      state=state)
+        else:
+            await message.delete()
+            msg1 = await message.answer(text='Идет загрузка данных...')
+            msg = await message.answer(text=ResultPrint(user_sn).today_button()[0],
+                                       reply_markup=main_kb)
+            await state.update_data(whitch_kb_was='main_kb')
+            await msg1.delete()
+            await support_function.delete_pre_message.del_pre_message(chat_id=msg.chat.id,
+                                                                      message_id=msg.message_id,
+                                                                      state=state)
